[REDIS] Java, Kotlin 캐시 추가

+++ kotlin/db/jpa/kotlin-jpa-mysql/src/main/kotlin/com/spring/jpa/mysql/repository/BoardRepository.kt
@@ -3,5 +3,5 @@ package com.spring.jpa.mysql.repository
 import com.spring.jpa.mysql.domain.Board
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface BoardRepository : JpaRepository<Board, Int> {
+interface BoardRepository : JpaRepository<Board, Long> {
 }
